Improved server config script to exclude certain countries and download maximum 20 configs per country only

+++ php/make_serverconfig.php
@@ -2,7 +2,7 @@
 
 ########################################################################
 #                                                                      #
-#   Serverconfig conversion script v4.01                               #
+#   Serverconfig conversion script v5.01                               #
 #   (c) by Tobse (cthullu@protonmail.com) in 2017                      #
 #                                                                      #
 #   This script converts NordVPN OpenVPN files in the ovpn directory   #
@@ -17,6 +17,14 @@
 
 // script may run one hour maximum
 ini_set("max_execution_time", 3600);
+
+// ignore the following top level country domains, separated by comma
+$ignore = "at,bh,ca,cn,de,fr,ge,in,ir,mx,nz,ru,sg,sw,sy,tr,uk,us,vn,zw";
+$ignorecountries = explode(",", $ignore);
+
+// maximum servers per one country
+// (to keep list small enough to remain performant)
+$maxPerCountry = 20;
 
 // regular expression server configs must match to be parsed
 $fileMatchPattern = "%^[a-z-]{2,5}\d+?\.(?:udp|tcp)(\d+?)\.ovpn$%i";
@@ -35,21 +43,47 @@ out("Retrieved data for " . count($hits[1])
 out("");
 if (!file_exists("ovpn")) mkdir("ovpn");
 $hits = $hits[0];
-natcasesort($hits);
+shuffle($hits);
 $cnt = 0;
+$downloaded = array();
 foreach ($hits as $hit) {
   $cnt++;
   $url = str_replace('"', "", $hit);
   $hitParts = explode("/", $url);
   $filename = preg_replace("%\.nordvpn\.com%", "",
                            $hitParts[count($hitParts) - 1]);
-  if (!preg_match($fileMatchPattern, $filename)
+
+  // don't download configs from ignored countries
+  $ignore = false;
+  foreach ($ignorecountries as $country) {
+    if (preg_match("%" . $country . "%i", $filename)) {
+      $ignore = true;
+      break;
+    }
+  }
+
+  // don't download more than $maxPerCountry configs per country
+  $servername = strstr($filename, ".", true);
+  preg_match_all("%[a-z]{2}%i", $servername, $spHits);
+  foreach ($spHits[0] as $checkCountry) {
+    if (!array_key_exists($checkCountry, $downloaded)) {
+      $downloaded[$checkCountry] = 0;
+    }
+    if (count($downloaded[$checkCountry]) > $maxPerCountry) {
+      $ignore = true;
+      break;
+    }
+  }
+  if ($ignore || !preg_match($fileMatchPattern, $filename)
    || file_exists("ovpn/" . $filename)) continue;
 
   // download ovpn file
   file_put_contents("ovpn/" . $filename, file_get_contents($url));
   out("Downloaded server config " . $filename . " ("
     . sprintf("%0.2f", 100/count($hits)*$cnt) . "%).");
+  foreach ($spHits[0] as $downloadedCountry) {
+    $downloaded[$downloadedCountry]++;
+  }
 }
 
 // create "serverconfigs" directory
